fea(statushandler): Test
